Added method to store credit transaction in transactions list

+++ library/src/main/java/com/thoughtworks/step/bank/DebitTransaction.java
@@ -4,11 +4,11 @@ import java.util.Date;
 
 public class DebitTransaction extends Transaction {
 
-    protected DebitTransaction(Date date, float amount, String to) {
+    protected DebitTransaction(Date date, double amount, String to) {
         super(date, amount, to);
     }
 
-    public DebitTransaction(float amount, String name) {
+    public DebitTransaction(double amount, String name) {
         this(new Date(), amount, name);
     }
 }
